requirements.txt
numpy
ephem
astropy
matplotlib
astroquery
tqdm
scipy
pyyaml
rich
